+++ src/test/java/oop/RectangleTest.java
@@ -6,21 +6,21 @@ import static org.testng.Assert.assertEquals;
 
 public class RectangleTest {
 
-    int length = 2;
-    int width = 1;
+    double length = 2;
+    double width = 1;
     Rectangle rectangle = new Rectangle(length, width);
 
     @Test
     public void calculatePerimeterThenShouldReturnTheResults() {
-        int expectedResults = 6;
-        int result = rectangle.calculatePerimeter();
+        double expectedResults = 6;
+        double result = rectangle.calculatePerimeter();
         assertEquals(result, expectedResults);
     }
 
     @Test
     public void calculateAreaThenShouldReturnTheResults() {
-        int expectedResults = 2;
-        int result = rectangle.calculateArea();
+        double expectedResults = 2;
+        double result = rectangle.calculateArea();
         assertEquals(result, expectedResults);
     }
 }
